+++ src/Transactions/CreateTransaction/CreateTransactionModal.js
@@ -1,4 +1,5 @@
 import PropTypes from 'prop-types';
+import { useCallback } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';
 
@@ -6,6 +7,7 @@ import stripesFinalForm from '@folio/stripes/final-form';
 import {
   Button,
   Col,
+  IconButton,
   Modal,
   ModalFooter,
   Row,
@@ -28,15 +30,18 @@ import { validateAllocationAmount } from './utils';
 
 import css from './CreateTransactionModal.css';
 
+const defaultFundsOptions = [];
+
 const CreateTransactionModal = ({
   allocationType,
   budget,
   form,
   fundId,
   handleSubmit,
+  isFundsLoading,
   isLoading,
   onClose,
-  fundsOptions = [],
+  fundsOptions = defaultFundsOptions,
   title,
   values: formValues,
 }) => {
@@ -47,6 +52,7 @@ const CreateTransactionModal = ({
   const hasToFundIdProperty = 'toFundId' in formValues;
   const hasFromFundIdProperty = 'fromFundId' in formValues;
   const fundLabelId = allocationType ? 'ui-finance.fund' : '';
+  const { fromFundId, toFundId } = formValues;
 
   let isToFundVisible = true;
   let isFromFundVisible = true;
@@ -62,15 +68,13 @@ const CreateTransactionModal = ({
   const isConfirmButtonDisabled = isLoading || invalid;
 
   const optionsFrom = (
-    (!hasToFundIdProperty ||
-    formValues.toFundId === fundId)
+    (!hasToFundIdProperty || formValues.toFundId === fundId)
       ? fundsOptions
       : fundsOptions.filter(f => f.value === fundId)
   );
 
   const optionsTo = (
-    (!hasFromFundIdProperty ||
-    formValues.fromFundId === fundId)
+    (!hasFromFundIdProperty || formValues.fromFundId === fundId)
       ? fundsOptions
       : fundsOptions.filter(f => f.value === fundId)
   );
@@ -94,6 +98,13 @@ const CreateTransactionModal = ({
     </ModalFooter>
   );
 
+  const swapFunds = useCallback(() => {
+    form.batch(() => {
+      form.change('toFundId', fromFundId);
+      form.change('fromFundId', toFundId);
+    });
+  }, [form, fromFundId, toFundId]);
+
   return (
     <Modal
       id="add-transfer-modal"
@@ -109,21 +120,34 @@ const CreateTransactionModal = ({
           {isFromFundVisible && (
             <Col xs>
               <FieldSelectionFinal
+                key={fromFundId}
                 dataOptions={optionsFrom}
                 label={<FormattedMessage id={fundLabelId || 'ui-finance.transaction.from'} />}
                 name="fromFundId"
                 disabled={!!allocationType}
+                loading={isFundsLoading}
               />
             </Col>
+          )}
+
+          {isFromFundVisible && isToFundVisible && (
+            <IconButton
+              className={css.swap}
+              icon="replace"
+              onClick={swapFunds}
+              disabled={isFundsLoading}
+            />
           )}
 
           {isToFundVisible && (
             <Col xs>
               <FieldSelectionFinal
+                key={toFundId}
                 dataOptions={optionsTo}
                 label={<FormattedMessage id={fundLabelId || 'ui-finance.transaction.to'} />}
                 name="toFundId"
                 disabled={!!allocationType}
+                loading={isFundsLoading}
               />
             </Col>
           )}